+++ packages/block-library/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 -   Added support for `border-control` to `Quote Block` ([47537](https://github.com/WordPress/gutenberg/pull/47537))
 
+## 8.3.0 (2023-02-01)
+
 ## 8.2.0 (2023-01-11)
 
 ## 8.1.0 (2023-01-02)